refactor: 리뷰 등록 버튼 분리
[#480]
