+++ frontend/src/components/ui/StationDetailsWindow/reviews/cards/ReviewCreate.tsx
@@ -1,13 +1,12 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import { modalActions } from '@stores/layout/modalStore';
-
-import { useCreateReview } from '@hooks/tanstack-query/station-details/reviews/useCreateReview';
 
 import Box from '@common/Box';
 import ButtonNext from '@common/ButtonNext';
 import FlexBox from '@common/FlexBox';
 
+import ReviewCreateButton from '@ui/StationDetailsWindow/reviews/cards/ReviewCreateButton';
 import ContentField from '@ui/StationDetailsWindow/reviews/common/ContentField';
 import HeaderWithRating from '@ui/StationDetailsWindow/reviews/common/HeaderWithRating';
 
@@ -19,22 +18,9 @@ const ReviewCreate = ({ stationId }: ReviewCreateProps) => {
   const [isReviewCreateOpen, setIsReviewCreateOpen] = useState(false);
   const [stars, setStars] = useState(5);
   const [content, setContent] = useState('');
-  const { createReview, isCreateReviewLoading } = useCreateReview(stationId);
-
-  useEffect(() => {
-    if (!isCreateReviewLoading && isReviewCreateOpen) {
-      setIsReviewCreateOpen(false);
-    }
-  }, [isCreateReviewLoading]);
 
   const handleClickReviewCreateCloseButton = () => {
     modalActions.closeModal();
-  };
-
-  const handleClickReviewCreateButton = () => {
-    if (content.length >= 5 && content.length <= 100) {
-      createReview({ stationId, ratings: stars, content });
-    }
   };
 
   const handleClickReviewCreateOpenButton = () => {
@@ -51,25 +37,14 @@ const ReviewCreate = ({ stationId }: ReviewCreateProps) => {
       )}
       <FlexBox nowrap>
         {isReviewCreateOpen ? (
-          <>
-            <ButtonNext
-              variant="outlined"
-              color="error"
-              fullWidth
-              onClick={handleClickReviewCreateOpenButton}
-            >
-              리뷰 그만작성하기
-            </ButtonNext>
-            <ButtonNext
-              disabled={isCreateReviewLoading || content.length < 5 || content.length > 100}
-              variant="contained"
-              color="success"
-              fullWidth
-              onClick={handleClickReviewCreateButton}
-            >
-              {isCreateReviewLoading ? '처리중...' : '등록'}
-            </ButtonNext>
-          </>
+          <ReviewCreateButton
+            content={content}
+            stars={stars}
+            stationId={stationId}
+            isReviewCreateOpen={isReviewCreateOpen}
+            setIsReviewCreateOpen={setIsReviewCreateOpen}
+            handleClickReviewCreateOpenButton={handleClickReviewCreateOpenButton}
+          />
         ) : (
           <>
             <ButtonNext
